added more files, some file changes, more in comments
