+++ app/src/main/AndroidManifest.xml
@@ -23,7 +23,7 @@
         <activity android:name=".MainActivity" />
         <activity android:name=".UserProfile" />
         <activity android:name=".Stats" />
-        <activity android:name=".Settings"></activity>
+        <activity android:name=".SettingsOld"></activity>
         <activity android:name=".POSTandGET"></activity>
     </application>
