+++ app/src/main/java/com/example/martin/personalenergytrainer/retrofit/APIRelated.kt
@@ -1,4 +1,4 @@
-package com.example.martin.personalenergytrainer.retrofit;
+package com.example.martin.personalenergytrainer.retrofit
 
 import com.google.gson.GsonBuilder
 import com.google.gson.JsonObject
@@ -18,13 +18,13 @@ class APIRelated {
         fun greetUser(@Path("user") user: String): Call<ResponseBody>
 
         @Headers("Content-type: application/json")
-        @POST("/api/post-some-data")
+        @POST("/api/post_some_data")
         fun getVectors(@Body body: JsonObject): Call<ResponseBody>
     }
 
     companion object{
         private val retrofit = Retrofit.Builder()
-                .baseUrl("http://192.168.1.4") //change to our url
+                .baseUrl("http://192.168.1.4:5000") //change to our url
                 .addConverterFactory(GsonConverterFactory.create(GsonBuilder().create()))
                 .build()
