+++ app/src/main/res/layout/activity_settings.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorBackground"
-    tools:context="com.example.martin.personalenergytrainer.Settings">
+    tools:context="com.example.martin.personalenergytrainer.SettingsOld">
 
     <RelativeLayout
         android:layout_width="match_parent"
